adjust language and spelling
A few misspellings have been corrected (for example, 'role' was
misspelled as 'roll', perhaps due to the use of speech input), and in
some areas the language slightly adjusted to make it clearer.  Empty
lines have been added between headers and paragraphs to make it easier
to visually separate.  It seems that the maximum line length used is for
70 characters, although some lines did not follow this; they have been
corrected.

In the 'documents' file, the list of accessible formats has been
converted from a paragraph description into a proper list, so that more
can be added easily in the future.